+++ index.html
@@ -8,7 +8,7 @@
   <title>Кексобукинг</title>
 </head>
 <body>
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
   <main>
     <div class="promo">
       <h1 class="promo__title visually-hidden">Keksobooking. Кексы по соседству</h1>
@@ -135,7 +135,7 @@
           <select id="capacity" name="capacity">
             <option value="3" selected>для 3 гостей</option>
             <option value="2">для 2 гостей</option>
-            <option value="1">для 1 гостя</option>
+            <option value="1">для 1 гостя</option> 
             <option value="0">не для гостей</option>
           </select>
         </fieldset>
