+++ js/util.js
@@ -1,0 +1,17 @@
+//Функция, возвращающая случайное целое число из переданного диапазона включительно
+function getRandomPositiveInteger (a, b) {
+  const lower = Math.min(Math.abs(a), Math.abs(b));
+  const upper = Math.max(Math.abs(a), Math.abs(b));
+  const result = Math.random() * (upper - lower) + lower;
+  return Math.round(result);
+}
+
+//Функция, возвращающая случайное число с плавающей точкой из переданного диапазона включительно
+function getRandomPositiveFloat (a, b, digits) {
+  const lower = Math.min(Math.abs(a), Math.abs(b));
+  const upper = Math.max(Math.abs(a), Math.abs(b));
+  const result = Math.random() * (upper - lower) + lower;
+  return +result.toFixed(digits);
+}
+
+export {getRandomPositiveInteger, getRandomPositiveFloat};
